+++ CHANGELOG.md
@@ -67,3 +67,4 @@ instance you've created and share it around with friends to simply import into t
 - Remove pack tags and add pack for semi public packs
 - Fix main Forge library missing caused packs not able to redownload it
 - Pretty print all json files written by the launcher
+- Add sha1/size to older Forge version libraries on install